ajustes

+++ src/main/java/br/com/project/airport/gestaofinanceira/model/Despesa.java
@@ -4,6 +4,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 
@@ -54,7 +56,7 @@ public class Despesa {
         return valor;
     }
 
-    public void setValor(Double valor) {
+    public void setValor(BigDecimal valor) {
         this.valor = valor;
     }
